Break up Zipfulldata worker into service classes
